requirements.txt
Django>=4.2,<5.3
djangorestframework>=3.14
django-cors-headers>=4.0
python-dotenv>=1.0
requests>=2.31
beautifulsoup4>=4.12
google-generativeai>=0.5
dj-database-url>=2.0
whitenoise>=6.6
gunicorn>=21.0
